NEW CONTRIB1.3 dependent modifications (mostly XERCES 3.0) new install instructions (install-macs.doxygen still needs update)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.3@4440 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/KERNEL/ConsensusMap.h
@@ -302,9 +302,9 @@ namespace OpenMS
 				std::partial_sort( tmp.begin(), tmp.begin()+n, tmp.end(), reverseComparator(Peak2D::IntensityLess()) );
 				for (Size element_index = 0; element_index < n; ++element_index )
 				{
-					output_map.push_back( ConsensusFeature( input_map_index, element_index, tmp[element_index] ) );
+					output_map.push_back( ConsensusFeature( (UInt)input_map_index, (UInt)element_index, tmp[element_index] ) );
 				}
-				output_map.getFileDescriptions()[input_map_index].size = n;
+				output_map.getFileDescriptions()[(UInt) input_map_index].size = (UInt)n;
 				output_map.updateRanges();
 				return;
 			}
